Adicionando a funcionalidade de Aluguel e validações

+++ src/main/java/com/javatechie/spring/mongo/api/service/LivroService.java
@@ -1,4 +1,4 @@
-package com.javatechie.spring.mongo.api.Service;
+package com.javatechie.spring.mongo.api.service;
 
 import com.javatechie.spring.mongo.api.model.Livro;
 import com.javatechie.spring.mongo.api.repository.LivroRepository;
@@ -31,11 +31,22 @@ public class LivroService {
         repository.save(livro);
     }
 
-    public void editarLivro(Livro livro) {
+    public String editarLivro(Livro livro) {
+        List<Livro> listaLivrosAlugados = repository.findByAlugadoTrue();
+        if(!listaLivrosAlugados.isEmpty()){
+            if(listaLivrosAlugados.contains(livro)){
+                return "O livro encontra-se alugado, por gentileza edite após devolver";
+            }
+        }
         repository.save(livro);
+        return "O livro foi editado com sucesso";
     }
 
     public String deleteById(int id) {
+        Optional<Livro> livroAlugado = repository.findByAlugadoTrueById();
+        if(livroAlugado.isPresent()){
+            return "O livro encontra-se alugado, por gentileza delete após devolver";
+        }
         if(repository.existsById(id)){
             repository.deleteById(id);
             return "Livro com o id " + id + " excluido com sucesso";
